benchmark poroelastic deviated wellbore
